# frontend/openapi/schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "WMS-SAEP API",
        "version": "0.1.0",
        "description": "API do Sistema de Requisição de Materiais",
        "contact": {
            "email": "contato@saep.local"
        }
    },
    "paths": {
        "/api/v1/auth/csrf/": {
            "get": {
                "operationId": "auth_csrf",
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CsrfTokenOutput"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/login/": {
            "post": {
                "operationId": "auth_login",
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginInput"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginInput"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginInput"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthSessionOutput"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/logout/": {
            "post": {
                "operationId": "auth_logout",
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/auth/me/": {
            "get": {
                "operationId": "auth_me",
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "sessionAuth401": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthSessionOutput"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/users/beneficiary-lookup/": {
            "get": {
                "operationId": "users_beneficiary_lookup",
                "parameters": [
                    {
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "sessionAuth401": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BeneficiaryLookupOutput"
                                    }
                                }
                            }
                        },
                        "description": "Lista curta de beneficiários elegíveis, limitada a 10 resultados."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AuthLoginInput": {
                "type": "object",
                "properties": {
                    "matricula_funcional": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    }
                },
                "required": [
                    "matricula_funcional",
                    "password"
                ]
            },
            "AuthSessionOutput": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "matricula_funcional": {
                        "type": "string",
                        "readOnly": true
                    },
                    "nome_completo": {
                        "type": "string",
                        "readOnly": true
                    },
                    "papel": {
                        "type": "string",
                        "readOnly": true
                    },
                    "setor": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AuthSetorOutput"
                            }
                        ],
                        "readOnly": true,
                        "nullable": true
                    },
                    "is_authenticated": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "is_authenticated",
                    "matricula_funcional",
                    "nome_completo",
                    "papel",
                    "setor"
                ]
            },
            "AuthSetorOutput": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "nome": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "nome"
                ]
            },
            "BeneficiaryLookupOutput": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "matricula_funcional": {
                        "type": "string",
                        "readOnly": true
                    },
                    "nome_completo": {
                        "type": "string",
                        "readOnly": true
                    },
                    "setor": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AuthSetorOutput"
                            }
                        ],
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "matricula_funcional",
                    "nome_completo",
                    "setor"
                ]
            },
            "CsrfTokenOutput": {
                "type": "object",
                "properties": {
                    "csrf_token": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "csrf_token"
                ]
            },
            "ErrorDetail": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "nullable": true
                    },
                    "trace_id": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "code",
                    "message"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                },
                "required": [
                    "error"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "sessionAuth401": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}